requirements.txt
Flask==3.0.0
tavily-python==0.3.0
flask_cors
